+++ kinto/core/resource/__init__.py
@@ -1095,7 +1095,7 @@ class Resource:
                         send_alert(self.request, message, url)
                     operator = COMPARISON.LT
 
-                if value == "" or not isinstance(value, (int, str)):
+                if value == "" or not isinstance(value, (int, str, type(None))):
                     raise_invalid(self.request, **error_details)
 
                 filters.append(Filter(self.model.modified_field, value, operator))